pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project.urls]
repository = "https://github.com/epicosy/sator"
homepage = "https://github.com/epicosy/sator"

[project]
name="sator"
version = "0.0.1"
description = "vulnerability database api"
readme = "README.md"
authors = [
    {name = "Eduard Pinconschi", email="eduard.pinconschi@tecnico.ulisboa.pt"},
    {name = "Sofia Reis", email="sofia.o.reis@tecnico.ulisboa.pt"}
]
keywords = ["security", "vulnerability", "database"]
license = {file = "LICENSE.txt"}
requires-python = ">=3.9"
classifiers = [
    "Topic :: Software Development",
    "Programming Language :: Python",
    "Environment :: Console",
    "Intended Audience :: Developers"
]
dependencies = [
    "cement==3.0.8",
    "colorlog==6.7.0",
    "Flask==2.2.2",
    "Flask-Cors==3.0.10",
    "Flask-GraphQL==2.0.1",
    "Flask-SQLAlchemy==3.0.2",
    "SQLAlchemy-Utils==0.39.0",
    "graphene-sqlalchemy==2.3.0",
    "graphene==2.1.9",
    "psycopg2==2.9.5",
    "pandas~=1.3.3",
    "PyYAML==6.0",
    "requests==2.28.2",
    "tqdm==4.64.1"
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "coverage",
    "twine>=1.11.0",
    "setuptools>=38.6.0",   
    "wheel>=0.31.0"
]


[tool.setuptools.packages.find]
exclude = ["ez_setup", "tests*"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
sator = ["tables/*"]

[project.scripts]
sator = "sator.main:main"
